fix(engine): sanitize contact-card vCards via a shared builder
The whatsapp-web.js adapter built a contact vCard by interpolating the raw
contact name and number, so CR/LF in either could inject extra vCard
fields/lines and a non-digit number corrupted the waid. The Baileys adapter
already sanitized this in a private method. Extract the sanitizing builder
(strip CR/LF, digits-only waid) into a shared module and use it from both
adapters, removing the duplicate.
